Make rand crate can be opt-out
