+++ Cargo.toml
@@ -12,4 +12,8 @@ repository = "https://github.com/jimmycuadra/retry"
 version = "1.1.0"
 
 [dependencies]
-rand = "0.7.3"
+rand = { version = "0.7.3", optional = true }
+
+[features]
+default = ["random"]
+random = ["rand"]
